requirements.txt
flask==3.1.0
tensorflow==2.17.1
numpy==1.26.4
pillow==11.0.0
google-cloud-storage==2.18.2
gunicorn==23.0.0